subtitleai/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic-extra-types==2.1.0
email-validator==2.1.0
celery==5.3.6
redis==5.0.1
fastapi-cache2==0.2.1
google-generativeai==0.3.2
openai-whisper==20231117
supabase==2.3.0
boto3==1.34.34
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
razorpay==1.4.1
sentry-sdk==1.40.0
python-ffmpeg==2.0.12
ffmpeg-python==0.2.0
moviepy==1.0.3
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
alembic==1.13.1
flower==2.0.1
structlog==23.2.0
python-json-logger==2.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
httpx==0.24.1
faker==22.0.0
bleach==6.1.0